fix: find all exact matches for reserved file names
Up until now, the regex only matched files with an extension,
(e.g. NUL.txt), which is probably the less common mistake.
Furthermore, whenever a file ended with a reserved name (e.g. COMICCON.txt),
it could not be added either.
I came up with a regex that solves these problems and is closer to the comments
on the StackOverflow thread that the script refers to.